requirements.txt

streamlit
openai
pillow
pytesseract
